Remove unnecessary boxing

+++ src/main/org/epics/archiverappliance/mgmt/archivepv/CapacityPlanningBPL.java
@@ -73,7 +73,7 @@ public class CapacityPlanningBPL {
                     int partitionSecond =
                             tempETLSource.getPartitionGranularity().getApproxSecondsPerChunk();
                     String identifyTmep88 = ((StorageMetrics) tempDest).getName();
-                    dataStoresAddingpv.put(identifyTmep88, Integer.valueOf(partitionSecond));
+                    dataStoresAddingpv.put(identifyTmep88, partitionSecond);
                     if (isDebug) logger.error(identifyTmep88 + "is added into dataStoresAddingpv");
                 }
             }
@@ -366,12 +366,12 @@ public class CapacityPlanningBPL {
                         if (TempValue == null) {
                             averagePercentageETL.put(
                                     identifypvAdding66,
-                                    Double.valueOf(tempETLMetrics33.estimateETLtimePercentageAfterPVadded));
+                                tempETLMetrics33.estimateETLtimePercentageAfterPVadded);
                             if (isDebug) logger.error(identifypvAdding66 + " is null ");
                         } else {
                             averagePercentageETL.put(
                                     identifypvAdding66,
-                                    Double.valueOf(TempValue + tempETLMetrics33.estimateETLtimePercentageAfterPVadded));
+                                TempValue + tempETLMetrics33.estimateETLtimePercentageAfterPVadded);
                             if (isDebug) logger.error(identifypvAdding66 + " is added ");
                         }
                     }
@@ -397,7 +397,7 @@ public class CapacityPlanningBPL {
                 String identifypvAdding6666 = entryTemp99666.getKey();
                 Double tempPercentageETL = entryTemp99666.getValue();
                 averagePercentageETL.put(
-                        identifypvAdding6666, Double.valueOf(tempPercentageETL / availableAppliancesNum));
+                        identifypvAdding6666, tempPercentageETL / availableAppliancesNum);
                 allNormalizationFactor.add(
                         new NormalizationFactor(identifypvAdding6666, tempPercentageETL.floatValue()));
             }
